+++ tests/Database/Factory/UnitTestingFactoryTest.php
@@ -491,7 +491,7 @@ class UnitTestingFactoryTest extends FrameworkTestCase {
 		];
 	}
 
-	public function test_custom_post_type() {
+	public function test_custom_post_type_for() {
 		register_post_type( 'custom-post' );
 
 		$post = static::factory()->post->for( 'custom-post' )->create_and_get();
@@ -536,6 +536,20 @@ class UnitTestingFactoryTest extends FrameworkTestCase {
 		$this->expectExceptionMessage( 'Error creating dynamic factory for conflict. Both post type and taxonomy exist.' );
 
 		static::factory()->conflict->create_and_get();
+	}
+
+	public function test_pass_factory_as_meta_argument() {
+		$post = static::factory()->post->with_meta( [
+			__FUNCTION__ => static::factory()->post->with_meta( 'example', 'true' ),
+		] )->create_and_get();
+
+		$meta = get_post_meta( $post->ID, __FUNCTION__, true );
+		$this->assertIsNumeric( $meta );
+
+		$underlying_post = get_post( $meta );
+
+		$this->assertInstanceOf( \WP_Post::class, $underlying_post );
+		$this->assertEquals( 'true', get_post_meta( $underlying_post->ID, 'example', true ) );
 	}
 
 	public static function slug_id_dataprovider(): array {